pipeline/requirements.txt
anthropic>=0.40.0
praw>=7.8.0
pytrends>=4.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
ruff>=0.8.0
